fix: Panic when Ident has no Obj
For some reason when the Ident is in another file in the
same package, the ast package does not provide an obj.
Avoid a panic that happens when trying to access the nil Obj.

This means the value cannot be tested for an exec, but at least
execinquery no longer crashes.